use example.org in some more places
example.org is already used in most places as default domain name.

* Use example.org instead of real (not reserved) domain names in some more
  places.

Using example.org is already much better than using real domain names as it was
done in the past, but still it has some negative effects.
If a dCache instance really uses them, the example.org server will get
“attacked”.
Or it could be even a security problem, when someone manages to actually run a
services on example.org, which dCache tries to use then (and potentially
trusts).

The best way IMHO would be to mark all options where an example value is given
with some flag, that they _must_ be overridden.
Another alternative, that should work in practise is using a domain like
“example.invalid” as this should already be sorted out by the local nameservers.
And ”invalid” is a specially reserved (top level) domain.

Signed-off-by: Christoph Anton Mitterer <mail@christoph.anton.mitterer.name>