dcache: billing -- fix error in liquibase varchar changeset
module: dcache

The liquibase changeset 4.1.7 was discovered to contain an off-by-one error in the string used to test the condition for changing the varchar length in the billing tables (the actual length of the string was 256, not 257). As a result, 4.1.7. 4.1.8 and 4.1.9 never will run, and all billing installations created by liquibase (which hardcodes the default to 256) remained as such.

The reason for the requested change was that the path and errormessage attributes of certain tables can well exceed 256.

The new changesets correct the error.

Target: master
Committed: master@db5ddfd6566edcbffa0829f16314bc9c2f0a3da8
Patch: http://rb.dcache.org/r/5555
Require-book: no
Require-notes: yes
Request: 2.2
Request: 2.6
Acked-by: Dmitry

RELEASE	NOTES: 	Fixes incorrect	sizes on billing table varchar attributes pnfsid, path and errormessage.  The problem arises with the billing databases	that were generated by Liquibase and not by hand. This is because Liquibase hardcodes the varchar default	to 256,	where postgresql leaves	it undefined (unlimited).
A previous patch to correct this was discovered to contain a bug which prevented it from running.

TESTING:

see patch http://rb.dcache.org/r/5555 for master