webadmin: add range to alarm query api and panel
module: webadmin

The webadmin alarm page is currently set to display a maximum of 100 entries before creating a new page.  However, there is no way at this time to limit the number of entries to be returned from the database.  This feature would of course be useful, especially when the service is set to store lower level log events.

This patch slightly modifies the underlying API for the query and adds two range fields to the filter object, which are then used to create the DataNucleus layer query.  The query panel now has an addition line where the user can optionally express a single range (0 to 10, 20 to 30, etc) of results to be displayed.

An implementation note.  I would have liked to make these text boxes spinners rather than simply typed for Integer.  However, the 1.5.9 implementation of wicketstuff-minis which provides a spinner behavior seems to be buggy (I could not get it to work).  There is a wicket-ui-jquery  library which also provides the spinner capability but it requires Wicket 6.0+.  I attempted to upgrade to 6.0 but encountered issues in the way that the new Wicket handles css's which was screwing up both table size and the javascript on some pages, so I left this aside.  I do think it a good idea eventually to move to 6.7, since we are somewhat lagging wrt Wicket development -- that is, if we are convinced that our wagon is hitched to Wicket on a long term basis.

Target: master
Committed: master@a43093ddecc4e79d7c42df288fb81d3bd5d3b066S
Patch: http://rb.dcache.org/r/5554
Acked-by: Dmitry
Require-notes: yes
Require-book: no
Request: 2.6
Merge-req:

RELEASE NOTES:

The webadmin alarms page now allows the user to express a single result range for the query (for instance, 0 to 100, 50 to 75, etc.).