Use correct junit dependency
It used to be the case that junit contained the Hamcrest classes rather
than having a Maven dependency on the Hamcrest artifact. To avoid
conflicts, junit provided the junit-dep artifact.

Since junit 4.11, Hamcrest is now specified as a dependency and it
is no longer necessary to use junit-dep.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/5593/