dcache-webadmin: Rename to webadmin.war in assembly
The httpd cell hard-codes the webapp name to the alias name, meaning
the war file must have the same name as the mount point. This is
obviously a bad coupling and it broke when we renamed the webadmin
module to dcache-webadmin, however in the interest of fixing this
regression quikly, this patch renames the artifact during packaging.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/5597/