Simplify use of DelayedReply
DelayedReply is used to provide message replies asynchronously. The existing
code blocks in the send() method until the cells code has injected the
endpoint used to send the reply. This forces the sender to catch an unlikely
InterruptedException and leads to deadlocks if the send method happens to be
called from the cells thread (as the injection will not happen until that
thread is released).

Both problems had already been addressed in the MessageReply class, however
the simpler DelayedReply class was never updated.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/5599/