webdav: Use HTTP for proxied upload
The webdav door allows uploads and downloads to be proxied throught the door
as an alternative to redirecting the client to the pool. For download we have
used HTTP for the transfer between the door and the pool, but for upload we
used an FTP data channel. We did this because originally the HTTP mover did
not support upload.

The current mover has however supported upload for quite a while and thus
it is time to let the door use HTTP for uploads too. The primary admin visible
change is that the TCP connection is now created from the door to the pool
rather than from the pool to the door. The configuration property
webdavInternalAddress is now used as a hint for the pool to select an interface
facing the door. The property is respected by both proxied downloads and
proxied uploads.

Target: trunk
Require-book: no
Require-notes: yes
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/5645/