Fix unit-tests for maven-2
Gerd reported that a recent patch (a019dc43) broke unit-tests.  This
was surprising as, before submitting, the patch was tested and all
unit-tests passed.

The patch introduced new dependencies, including
'powermock-api-mockito'.  This dependency, in turn, declares a
dependency on mockito-all, which is the Mockito jar that includes all
its dependencies within the jar file.  One of these dependencies is
Hamcrest; i.e., the mockito-all jar contains class files from some
unspecified (but probably v1.1) version of Hamcrest.

Unit tests were compiled against the declared version of Hamcrest
(v1.3).  When run within the JUnit framework, the classes loaded will
depend on the classpath order.  When mockito-all precedes the Hamcrest
jar, some unit tests were in error.  These are tests that use methods
in Hamcrest that have a different signature between v1.1 and v1.3.
Therefore, these tests would throw a NoSuchMethodException when run.

It seems that, for Maven v3, the Hamcrest v1.3 jar appeared before the
mockito-all jar whereas, for Maven v2, it is the other way around.
Therefore, the patch worked when tested, but only by chance.

This patch fixes this problem by excluding the mockito-all dependency
and including the mockito-core dependency.  The Mockito classes in the
latter has the same dependencies as the former, but this is expressed
through the pom.xml file.  For both compilation and running within
JUnit, the Hamcrest dependency is satisfied by the Hamcrest v1.3 jar
file.

Target: trunk
Requires-notes: no
Requires-book: no
Acked-by: Gerd Behrmann
Patch: http://rb.dcache.org/r/5704/